Stop deploying `mozc_tip64.dll.pdb` again (#1262)
This commit effectively reverts my previous commit [1], which aimed to
deploy the PDB file for 'mozc_tip64.dll' for better debuggability.

It turns out that 'clang-cl' does not support the '/PDBSTRIP' option,
which means that a further change is needed to keep the installer size
as similar as possible when switching from GYP to Bazel.

Let's undo this change so as not to block the Bazel migration.

Closes #1261

 [1]: 2a529d57b814b2618f8d8370246df1aefb08e92a

PiperOrigin-RevId: 749060706