Override `QT_QPA_PLATFORM` only on Linux desktop (#1271)
This follows up to my previous commit [1], which attempted to let Qt
mozc_renderer always use X11 protocol to work around #794.

What is remaining is that we want to keep Qt mozc_renderer buildable
not only on Linux but also on macOS and Windows. So let's apply the
above workaround only for Linux desktop.

 [1]: a31dd834f24433cd508c6cf16b3212778be44116

PiperOrigin-RevId: 756115427