Make embedded preedit TSVs consistent (#1282)
With this commit the following three preedit TSVs are embedded not only
in the executabled built with Bazel but also executables built with GYP.

 * src/data/preedit/50keys-hiragana.tsv
 * src/data/preedit/flick-number.tsv
 * src/data/preedit/toggle_flick-number.tsv

Note that the above missing TSV files are used only in the mobile
environment. Basically there should be no user-observable changes except
for the binary diff between Bazel build and GYP build becomes smaller.

Here are the file size diffs in the GYP build.

    File Name        Old       New    Diff
 ------------------------------------------
 mozc_broker.exe   1966592   1976320   9728
 mozc_server.exe  19974144  19984384  10240
 mozc_tip32.dll    4952576   4962304   9728
 mozc_tip64.dll    5143552   5153280   9728
 mozc_tool.exe     2954240   2963968   9728

Ideally the diff should be observed only in the executable where the
embedded file is actually used, but it is a different story that needs
to be optimized in the future.

Closes #1281.

PiperOrigin-RevId: 757038357