Stop calling `CrashReportHandler::Initialize` (#1285)
To follow up my previous commit [1], with this commit we stop calling
CrashReportHandler::Initialize in all the remaining executables.

This is a major milestone towards removing CrashReportHandler and the
dependency on breakpad (#1263).

Note that there must be no observable behavior change in the OSS build,
where CrashReportHandler has been no-op. It has been funtional when and
only when 1) 'branding' is 'GoogleJapaneseInput' and 2) executables are
built with GYP.

CrashReportHandler and the dependency on breakpad are going to be
removed in subsequent commits.

 [1]: 5045e397556144ac8e5dd0b6f1716ef9358c8702

PiperOrigin-RevId: 760444827