Share platforms among rules in Windows bazel build (#1297)
This reworks my precious commit [1], which introduced our initial
support of CC toolchain resolution as planned in #1112.

While all the executables are still built as expected, defining a local
private platform for every build target is turned out to be quite
inefficient, as Bazel build their build dependencies separately as if
they do not share the same CPU architecture.

    platform_name = "_" + name + "_platform"
    native.platform(
        name = platform_name,
        constraint_values = [
            cpu,
            "@platforms//os:windows",
        ],
        visibility = ["//visibility:private"],
    )

A simple and actually well-known solution is to define platforms
globally then always refer to them.

  platform(
      name = "windows-x86_32",
      constraint_values = [
          "@platforms//cpu:x86_32",
          "@platforms//os:windows",
      ],
  )

  platform(
      name = "windows-x86_64",
      constraint_values = [
          "@platforms//cpu:x86_64",
          "@platforms//os:windows",
      ],
  )

The other motivation of this change is to start relying on --platforms
commandline option to specify the target CPU architecture when building
Mozc64.msi for ARM64. Using --platforms option for this purpose is
indeed a well known approach in Bazel, and following the same convention
probably makes developers' lives easier.

In this commit we still specify

  --platforms=//:windows-x86_64

in .bazelrc as Mozc64.msi is not yet compatible with ARM64, but
eventually we can use the host CPU architecture as the default target
CPU architecture with allowing developers to override it at the command
line.

To summarize:

 * no behavior change in the final artifacts.
 * much faster build, which is almost the same as GYP build.
   * Number of build targets: 8,109 -> 4,457
   * Build time (bazel part): 50 min -> 31 min on GitHub Actions

Closes #1295.

 [1]: 2766b7d29aa917ae4c0066515843e08c02d9181e

PiperOrigin-RevId: 764986202