Build `Mozc64.msi` with Bazel in Windows CI (#1303)
This commit removes the `_bazel` suffix from the Windows CI jobs and
artifacts, and adds `_gyp` suffix instead to GYP-based ones.

We have put '_bazel' suffix in GitHub Actions for Windows to make it
clear that jobs and artifacts with Bazel are still experimental [1]. Now
that Bazel-based build is becoming production-ready (#1272) and the GYP
build is being deprecated, let's put '_gyp' suffix instead for GYP-based
jobs and artifacts.

This commit is only about how our CI jobs and artifacts are labeled.
There must be no user-visible change in the final artifacts.

 [1]: 9cf9d2d6f2ef83e5d1532be2759a9a525b9d2d29

PiperOrigin-RevId: 766550830