Make `absl.gyp` compatible with Abseil 20250512.0 (#1308)
This follows up to our previous commit [1], which updated abseil-cpp
version from 20250127.0 to 20250512.0 only for Bazel build.

Before applying the same change to GYP build, this commit ensures that
the 'base/absl.gyp' is compatible with both 20250127.0 and 20250512.0.

Until we actually update abseil-cpp version in GYP build, there must be
no change in the final artifacts.

 [1]: f2ce27b8cb148b63dc78a23b30b619119d831ca5

PiperOrigin-RevId: 768074353