Start building `Mozc64.msi` for ARM64 environment (#1305)
This is an important milestone towards supporting ARM64 Windows (#1130).

With this commit, we can start building 'Mozc64.msi' that is compatible
with ARM64 processes and x86 processes on Windows. To build 'Mozc64.msi'
to be ARM64 compatible, run the following commands:

  python build_tools/build_qt.py --release --confirm_license          \
     --target_arch=arm64
  bazelisk build package --config oss_windows --config release_build  \
     --platforms=//:windows-arm64

Note that no additional Visual Studio components are required unless you
explicitly specify the above options. Existing continuous builds for x64
should continue to work without any changes.

Note also that using Mozc in x64 processes on ARM64 machines is not yet
supported, as it requires us to build an ARM64X DLL that can be loaded
into both x64 and arm64 processes. It will be handled in a subsequent
commit.

PiperOrigin-RevId: 769539948