bzlmod: Use `http_archive` to download `wil` (#1321)
As a preparation for stopping using git-submodules (#1320), this commit
switches the 'wil' dependency from a git submodule to a 'http_archive'
rule in the 'MODULE.bazel' file.

Note that the Wil v1.0.230629.1 is chosen to be the nearest tagged
version from the commit that is currently used [1]. We prefer to use the
source archive associated with a tag as a lesson learned from the past
incident [2].

In general there must be no observable changes in the final artifacts.

 [1]: https://github.com/microsoft/wil/commit/fc5dbf55989fe20351c71d038a8d12de4b397a6d
 [2]: https://github.blog/open-source/git/update-on-the-future-stability-of-source-code-archives-and-hashes/

PiperOrigin-RevId: 778848067