Inline `//:intel_cpu` for simplicity (#1323)
This follows up to my previous commit [1], which introduced a
synthesized config '//:intel_cpu' that is set when either
"@platforms//cpu:x86_32" or "@platforms//cpu:x86_64" is set as part of
the effort to speed up our Bazel-based build for Windows (#1295).

The problem is that in downstream targets we can still write select
rules without '//:intel_cpu' as follows:

  select({
    "@platforms//cpu:x86_32": ["/CETCOMPAT"],
    "@platforms//cpu:x86_64": ["/CETCOMPAT"],
    "//conditions:default": [],
  }),

Let's just write it be written as a redundant but more explicit select
rules for now.

 [1]: f60a12e3d2db498aa8dc717ecc629977abd5cf5f

PiperOrigin-RevId: 778849008