Support different interpolation weights for decoding and NWP.
Base scores of decoding and NWP are very different. In decoding, scores are
mostly assigned by the underlying class language model and it includes the
score of reading model. Taking interpolation between these base scores and
rescoring LM's scores is thus reasonable. However, in NWP, many components
provide results with rule-based scores just for ordering (e.g., 1, 10, 20, ...
etc.). Therefore, it is often meaningless to take interpolation with such
scores. By supporting different weights, one can control the behavior in a more
flexible manner.

This CL renames the proto field name but this is safe, as it is not referenced
yet even in canary.

PiperOrigin-RevId: 783650572