Deprecates convertwithnodeinfo ops.
Since Converter doesn't directly depend on the lattice, we want to get rid of lattice-specific ops from converter_main. We would like to move it to immutable_converter_main if necessary. Then, we can avoid tricky implementation based on global(singleton) flags as we have full access to the lattice.

PiperOrigin-RevId: 797106290