Deprecates bnext and enext linked list for the sake of simplicity. Uses the simple std::vector instead.
Since Lattice is now almost local storage, using linked list might not help the performance. Removed unused fields from Node.

This CL changes the node insertion order into the lattice, which may cause the final ranking to change when the costs are the same. However, the behavior when costs are identical is undefined in the first place, so this change does not fix that issue. We do not allow the reranking depending on the node insertion order.

Since the old personal-name re-segmentation puts the node with the same cost, it puts -1 to boost the re-segmented node.

PiperOrigin-RevId: 798991320