Rely on RAII around `SECURITY_ATTRIBUTES` (#1383)
This is a semi-mechanical refactoring to improve the error handling
around 'SECURITY_ATTRIBUTES' as a preparation to start using
'PROC_THREAD_ATTRIBUTE_JOB_LIST' (#1377).

The issue is that the current method signature of

  static bool WinSandbox::MakeSecurityAttributes(
      ObjectSecurityType shareble_object_type,
      SECURITY_ATTRIBUTES* security_attributes);

makes it difficult to ensure that

  ::LocalFree(security_attributes->lpSecurityDescriptor);

is always called. The idea here is to change that method signature to
directly return an RAII wrapper 'wil::unique_hlocal_security_descriptor'
instead of updating 'SECURITY_ATTRIBUTES' in-place.

  static wil::unique_hlocal_security_descriptor
  WinSandbox::MakeSecurityDescriptor(
      ObjectSecurityType shareble_object_type);

With that, the caller can simply create 'SECURITY_ATTRIBUTES' with
making it clear that 'wil::unique_hlocal_security_descriptor' takes care
of the cleanup as follows:

  auto security_descriptor =
      MakeSecurityDescriptor(WinSandbox::kSharablePipe);
  if (!security_descriptor) {
    LOG(ERROR) << "Cannot make SecurityDescriptor";
    return;
  }
  SECURITY_ATTRIBUTES security_attributes = {
      .nLength = sizeof(SECURITY_ATTRIBUTES),
      .lpSecurityDescriptor = security_descriptor.get(),
      .bInheritHandle = FALSE,
  };
  auto handle = ::CreateNamedPipe(..., &security_attributes);
  if (handle == INVALID_HANDLE_VALUE) {
    return;
  }

There must be no behavior change in the successful paths. For error
cases, the new behavior should be more appropriate as 'LocalFree' is now
guaranteed to be called.

PiperOrigin-RevId: 820648748