Apply the suggestion filter to next word predictions.
In the next word prediction, both `request_key_` and `result.key` are empty
and, hence, it is detected as the exact match case unexpectedly. This CL not
only fixes this issue but also introduces a stricter filtering rule for the
next word prediction. Now the filter is applied both for the result value and
for "history + result.value" to suppress the case where the concatenation of
the history and predicted value becomes an offensive word.

PiperOrigin-RevId: 822351798