Use `PROC_THREAD_ATTRIBUTE_JOB_LIST` to set a job (#1397)
On Windows 10 and later versions, PROC_THREAD_ATTRIBUTE_JOB_LIST
attribute is a more recommended way to assign a process to job objects
when creating a process than calling AssignProcessToJobObject API after
the process creation [1]. Let's use it here to modernize the code.

One caveat is that we still need to create the process in a suspended
state so that we can call SetThreadToken API before the main thread
starts running. Several comments are added to make it clear.

Closes #1377.

 [1]: https://devblogs.microsoft.com/oldnewthing/20230209-00/?p=107812

PiperOrigin-RevId: 829711929