Make `<RegistrySearch>` work in the repair mode (#1411)
It turns out that <Property> with <RegistrySearch> does not work in the
repair/uninstall mode unless the property is marked as Secure. As a
result, if the user tries to install the same version of Mozc again,
then they would see a misleading error messages (#1410).

Affected conditions are:
 * ARM64 check [1]
 * Windows build number check [2]

This commit marks the relevant properties as Secure to fix the issue.

 [1]: 554848a9780a5bbaa009dabce11208d0710e71e3
 [2]: f5b4ce9062b8a96ef685a271674e6aa0d78523a8

Closes #1410.

PiperOrigin-RevId: 850031011